Create tip calculator code
This code calculates the bill of each person in the group at a restaurant. It asks the total bill of group members. Then it asks the percentage of tip the group members want to pay. Then it asks the total number of group members. Finally it displays the final amount of bill to be payed by each of the group member.